Fixed RMWAbstract and ConstantPropagation referencing Linux tags instead of the general tags. Moved tso.Tag as subclass into Tag
