--- dartagnan/src/main/java/com/dat3m/dartagnan/program/event/arch/tso/utils/Tag.java
@@ -1,6 +1,0 @@
-package com.dat3m.dartagnan.program.event.arch.tso.utils;
-
-public class Tag extends com.dat3m.dartagnan.program.event.Tag {
-
-    public static final String ATOM      = "A";
-}
